combined cgram and color_math
added Color struct, implemented direct color
